src/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-multipart
requests
